split employee to pages
