+++ frontend/src/employee-frontend/App.tsx
@@ -44,8 +44,8 @@ import { sessionKeepalive } from './components/common/sessionKeepalive'
 import DecisionPage from './components/decision-draft/DecisionDraft'
 import DocumentTemplatesPage from './components/document-templates/template-editor/DocumentTemplatesPage'
 import TemplateEditorPage from './components/document-templates/template-editor/TemplateEditorPage'
-import EmployeePinCodePage from './components/employee/EmployeePinCodePage'
-import EmployeePreferredFirstNamePage from './components/employee/EmployeePreferredFirstNamePage'
+import EmployeePinCodePage from './components/employee-pin-code/EmployeePinCodePage'
+import EmployeePreferredFirstNamePage from './components/employee-preferred-first-name/EmployeePreferredFirstNamePage'
 import EmployeePage from './components/employees/EmployeePage'
 import EmployeesPage from './components/employees/EmployeesPage'
 import FeeDecisionDetailsPage from './components/fee-decision-details/FeeDecisionDetailsPage'
