+++ frontend/src/employee-frontend/components/employee-preferred-first-name/queries.ts
@@ -6,9 +6,7 @@ import { Queries } from 'lib-common/query'
 
 import {
   getEmployeePreferredFirstName,
-  isPinLocked,
-  setEmployeePreferredFirstName,
-  upsertPinCode
+  setEmployeePreferredFirstName
 } from '../../generated/api-clients/pis'
 
 const q = new Queries()
@@ -21,9 +19,3 @@ export const setEmployeePreferredFirstNameMutation = q.mutation(
   setEmployeePreferredFirstName,
   [employeePreferredFirstNameQuery]
 )
-
-export const isPinLockedQuery = q.query(isPinLocked)
-
-export const upsertPinCodeMutation = q.mutation(upsertPinCode, [
-  isPinLockedQuery
-])
